Failed access to 绩点表 right now.
So I have to upload this as evidence to prove that I earned the required points, which is 174. 
By the way, I see the course from Teacher Ou Guangyu is still in the table, which the college academic affairs teacher promised me that they(she) would help me de-select the course.